+++ test/sourceParams.js
@@ -114,6 +114,47 @@ describe('sourceParams', () => {
       }
     })
 
+    it('should set params from rooseveltConfig.json', () => {
+      // set app directory
+      const appDir = path.join(__dirname, 'app/sourceParams')
+
+      // build roosevelt config from sample
+      const configJson = {
+        ...sampleConfig,
+        enableCLIFlags: false
+      }
+
+      // create app directory
+      fs.ensureDirSync(path.join(appDir))
+
+      // generate rooseveltConfig.json with sample config
+      fs.writeJSONSync(path.join(appDir, 'rooseveltConfig.json'), configJson)
+
+      // initialize roosevelt
+      const app = require('../roosevelt')({
+        appDir: appDir
+      })
+
+      const appConfig = app.expressApp.get('params')
+
+      // do some param post-processing that matches what we expect from roosevelt
+      configJson.staticsRoot = path.join(appDir, configJson.staticsRoot)
+      configJson.publicFolder = (path.join(appDir, configJson.publicFolder))
+      configJson.css.sourcePath = path.join(configJson.staticsRoot, configJson.css.sourcePath)
+      configJson.css.output = path.join(configJson.publicFolder, configJson.css.output)
+      configJson.js.sourcePath = path.join(configJson.staticsRoot, configJson.js.sourcePath)
+      configJson.clientViews.output = path.join(configJson.staticsRoot, configJson.clientViews.output)
+
+      // for each param, test that its value is set in roosevelt
+      for (const key in appConfig) {
+        const param = appConfig[key]
+
+        if (!blacklist.includes(key)) {
+          assert.deepStrictEqual(param, configJson[key], `${key} was not correctly set`)
+        }
+      }
+    })
+
     it('should resolve variables in params', () => {
       // build roosevelt config with lots of variables
       const config = {
@@ -130,7 +171,7 @@ describe('sourceParams', () => {
         https: {
           port: '${(port + 1)}' // eslint-disable-line
         },
-        multipart: {
+        formidable: {
           multiples: '${versionedPublic}' // eslint-disable-line
         },
         css: {
@@ -166,7 +207,7 @@ describe('sourceParams', () => {
       assert.deepStrictEqual(appConfig.css.whitelist[0], path.join(appConfig.staticsRoot, 'coolCss/hello.js'), 'partial param variable not parsed correctly')
       assert.deepStrictEqual(appConfig.staticsSymlinksToPublic[0], path.join(appConfig.staticsRoot, 'coolJavaScript'), 'param variable within array not parsed correctly')
       assert.deepStrictEqual(appConfig.js.webpack.bundles[0].output, path.join(appConfig.staticsRoot, 'coolCss/hello.js'), 'deeply nested param variable not parsed correctly')
-      assert.deepStrictEqual(appConfig.multipart.multiples, true, 'true param variable not parsed correctly')
+      assert.deepStrictEqual(appConfig.formidable.multiples, true, 'true param variable not parsed correctly')
       assert.deepStrictEqual(appConfig.alwaysHostPublic, false, 'false param variable not parsed correctly')
     })
   })